Edit failing test
The test test_asymmetric_replicator_dynamics_size_2_3_given_timepoints
was passing in windows and ubuntu machines but failed for Mac on GitHub.

Attempt 1: Try and fix the issue by increasing the max value and num
value of the timepoints array. Also more meaningful payoff matrices
are used and the check occurs just for the first and last entry of the
output of asymmetric replicator dynamics function.