Implement asymmetric replicator dynamics (#97)
* Add initial tests

Two property based tests added; one for the derivative and one for the
asymmetric_replicator_dynamics function.

* Implement asymmetric replicator dynamics

* Minor bug fixes on tests

* Minor changes and refactoring

* Run black

* Raise ValueError when A and B have unequal sizes

At the start of asymmetric_replicator_dynamics check if A and B have the
same dimensions and raise a ValueError if they don't.

* Add additional tests

- Test for get_derivative_of_asymmetric_fitness
- Property test of equivalence between symmetric and asymmetric method
- Example test using default values
- Example test by passing timepoints arguments
- Example test by passing x0 and y0 arguments
- Example of erroneous test when passing incorrect dimensions of A nad B

* Move dimensions check of matrices in Game class

Check that the shape of A and B are the same in the __init__ method of
the Game class. Remove it from replicator_dynamics.py and remove its
equivalent test from test_replicator_dynamics.py

* Fix minor typos and refactor some code

* Limit the matrices values in the equivalence test

Change the range of values of the payoff matrices in the equivalnce test
from 1 - 20 to 1 - 5. The test fails for some cases when the maximum is
at 20 possibly because of the way odeint works. The test might need to
be fixed in another way.

* Change the way dimensions are check

Couldn't use array.shape because nashpy also accepts lists of lists
as inputs. Used a more generic way to check dimensions that apply
both to arrays and lists.

* Add test for incorrect dimensions of A and B

* Minor changes to docstring and test

Use absolute tolerance parameter instead of relative tolerance parameter
in some tests.

Add some additional notes to one of the docstrings to explain why a
small hypothesis range is used.

* Edit failing test

The test test_asymmetric_replicator_dynamics_size_2_3_given_timepoints
was passing in windows and ubuntu machines but failed for Mac on GitHub.

Attempt 1: Try and fix the issue by increasing the max value and num
value of the timepoints array. Also more meaningful payoff matrices
are used and the check occurs just for the first and last entry of the
output of asymmetric replicator dynamics function.

* Add test for the game class

Add a test of the method of asymmetric_replicator_dynamics

* Add asymmetric_replicator_dynamics in Game class

* Write how-to section

* Add reference for asymmetric replicator dynamics

* Write reference section

Wirte the reference section explaining how the algorithm works and the
mathematics behind it.

* Minor changes in doctests

* Only include the first 5 decimals for the doctests

Change the format of the output and using ELLIPSIS' marker (...) only
consider the first 5 decimals of every entry of xs and ys

* Add plots for the docs

Add a plot for a convergent and a non-convergent game using
asymmetric replicator dynamics

* Add discussion section